ats-/backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.13.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
bcrypt==4.3.0
blis==0.7.11
boto3==1.40.6
botocore==1.40.6
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.6
cymem==2.0.11
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.3.0
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.20.0
fsspec==2025.9.0
googleapis-common-protos==1.71.0
grpcio-status==1.62.3
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.36.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.3.2
mpmath==1.3.0
msal==1.33.0
murmurhash==1.0.13
networkx==3.4.2
numpy==1.26.4
ollama==0.3.1
packaging==25.0
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.0.0
pluggy==1.6.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic-settings==2.1.0
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pypdfium2==5.0.0
pytesseract==0.3.13
pytest-asyncio==0.21.2
pytest==7.4.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.0
python-dotenv==1.1.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2025.10.23
requests==2.32.4
rich==14.1.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.4.2
scipy==1.15.3
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy==3.7.5
srsly==2.5.1
starlette==0.27.0
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==12.0
wrapt==1.17.2
